Added error counting
