+++ src/code/llvm/types.rs
@@ -45,49 +45,9 @@ impl<'c> Compiler<'c> {
         ast
     }
 
-    /// Generate code for a full function definition
-    pub fn gen_fundef(&mut self, proto: &FunProto, body: &Vec<AstPos>) {
-        if self.current_fn.is_some() {
-            panic!("Nested functions are not currently supported, function {} must be moved to the top level", proto.name);
-        }
-
-        let old_vars = self.vars.clone();
-
-        let f = match self.module.get_function(proto.name.as_str()) {
-            Some(f) => f,
-            None => self.gen_fun_proto(proto).unwrap(),
-        };
-        self.current_fn = Some(f);
-        self.current_proto = Some(proto.clone());
-
-        let bb = self.ctx.append_basic_block(f, "fn_entry"); //Add the first basic block
-        self.build.position_at_end(bb); //Start inserting into the function
-
-        //Add argument names to the list of variables we can use
-        for (arg, (ty, proto_arg)) in f.get_param_iter().zip(proto.args.iter()) {
-            let alloca = self.entry_alloca(
-                proto_arg.clone().unwrap_or("".to_owned()).as_str(),
-                self.llvm_type(ty),
-            );
-            self.build.build_store(alloca, arg); //Store the initial value in the function parameters
-
-            if let Some(name) = proto_arg {
-                self.vars.insert(name.clone(), (alloca, ty.clone()));
-            }
-        }
-
-        //Generate code for the function body
-        for ast in body {
-            self.gen(ast, false);
-        }
-
-        self.vars = old_vars; //Reset the variables
-        self.current_fn = None;
-        self.current_proto = None;
-    }
 
     /// Generate code for a function prototype
-    fn gen_fun_proto(&mut self, proto: &FunProto) -> Result<FunctionValue<'c>, String> {
+    pub(super) fn gen_fun_proto(&mut self, proto: &FunProto) -> Result<FunctionValue<'c>, String> {
         if self.module.get_function(proto.name.as_str()).is_some() {
             return Err(format!("Function {} defined twice", proto.name));
         }
